requirements.txt
paho-mqtt>=2.0.0
python-dotenv>=1.2.1
pyserial>=3.5